update config and launch files
