+++ waywiser/launch/carla_truck.launch.py
@@ -13,7 +13,6 @@ def generate_launch_description():
     waywiser_core_dir = get_package_share_directory('waywiser_core')
     waywiser_teleop_dir = get_package_share_directory('waywiser_teleop')
     waywiser_twist_safety_dir = get_package_share_directory('waywiser_twist_safety')
-    waywiser_hwbringup_dir = get_package_share_directory('waywiser_hwbringup')
 
     # args that can be set from the command line or a default will be used
     use_sim_time_la = DeclareLaunchArgument(
@@ -68,6 +67,11 @@ def generate_launch_description():
         'control_vehicle_node',
         default_value='waywiser_truck_node',
         description='Name of the vehicle node to control',
+    )
+    frame_prefix_la = DeclareLaunchArgument(
+        'frame_prefix',
+        default_value='/',
+        description='Prefix to publish robot transforms in',
     )
     localization_node_name_la = DeclareLaunchArgument(
         'localization_node_name',
@@ -126,19 +130,20 @@ def generate_launch_description():
         }.items(),
     )
 
-    waywiser_truck = IncludeLaunchDescription(
+    waywiser_truck_launch = IncludeLaunchDescription(
         PythonLaunchDescriptionSource(
             [
                 os.path.join(
-                    waywiser_hwbringup_dir,
+                    waywiser_core_dir,
                     'launch',
-                    'truck.launch.py',
+                    'waywiser_truck.launch.py',
                 )
             ]
         ),
         launch_arguments={
             'use_sim_time': LaunchConfiguration('use_sim_time'),
             'vehicle_config': LaunchConfiguration('vehicle_config'),
+            'frame_prefix': LaunchConfiguration('frame_prefix'),
         }.items(),
     )
 
@@ -155,22 +160,6 @@ def generate_launch_description():
         launch_arguments={
             'localization_config': LaunchConfiguration('vehicle_config'),
             'localization_node_name': LaunchConfiguration('localization_node_name'),
-        }.items(),
-    )
-
-    tfs_to_navsatfixfused = IncludeLaunchDescription(
-        PythonLaunchDescriptionSource(
-            [
-                os.path.join(
-                    waywiser_core_dir,
-                    'launch',
-                    'tfs_to_navsatfixfused.launch.py',
-                )
-            ]
-        ),
-        launch_arguments={
-            'use_sim_time': LaunchConfiguration('use_sim_time'),
-            'config': LaunchConfiguration('vehicle_config'),
         }.items(),
     )
 
@@ -242,14 +231,14 @@ def generate_launch_description():
     ld.add_action(rviz2_la)
     ld.add_action(control_vehicle_node_name_la)
     ld.add_action(localization_node_name_la)
+    ld.add_action(frame_prefix_la)
 
     # start nodes
     ld.add_action(carla_orchestrator)
     ld.add_action(waywiser_carla_relay)
     ld.add_action(twist_safety)
-    ld.add_action(waywiser_truck)
+    ld.add_action(waywiser_truck_launch)
     ld.add_action(waywiser_truck_localization_launch)
-    ld.add_action(tfs_to_navsatfixfused)
     ld.add_action(emulated_angle_sensor)
     ld.add_action(emulated_range_sensor_array)
     ld.add_action(teleop_rviz2)
