+++ waywiser/launch/hwbringup_truck.launch.py
@@ -2,12 +2,16 @@ import os
 
 from ament_index_python import get_package_share_directory
 from launch import LaunchDescription
-from launch.actions import DeclareLaunchArgument, IncludeLaunchDescription
+from launch.actions import DeclareLaunchArgument, IncludeLaunchDescription, OpaqueFunction
 from launch.launch_description_sources import PythonLaunchDescriptionSource
 from launch.substitutions import LaunchConfiguration
 
+from waywiser_py.waywiser_utils import get_full_file_path
+import yaml
+
 
 def generate_launch_description():
+    waywiser_core_dir = get_package_share_directory('waywiser_core')
     waywiser_hwbringup_dir = get_package_share_directory('waywiser_hwbringup')
     waywiser_twist_safety_dir = get_package_share_directory('waywiser_twist_safety')
     waywiser_teleop_dir = get_package_share_directory('waywiser_teleop')
@@ -48,20 +52,47 @@ def generate_launch_description():
         default_value='waywiser_truck_node',
         description='Name of the vehicle node to control',
     )
+    frame_prefix_la = DeclareLaunchArgument(
+        'frame_prefix',
+        default_value='/',
+        description='Prefix to publish robot transforms in',
+    )
+    localization_node_name_la = DeclareLaunchArgument(
+        'localization_node_name',
+        default_value='waywiser_truck_localization_node',
+        description='Name of the node to be launched',
+    )
 
     # include launch files
-    truck = IncludeLaunchDescription(
+    waywiser_truck_launch = IncludeLaunchDescription(
         PythonLaunchDescriptionSource(
             [
                 os.path.join(
-                    waywiser_hwbringup_dir,
+                    waywiser_core_dir,
                     'launch',
-                    'truck.launch.py',
+                    'waywiser_truck.launch.py',
                 )
             ]
         ),
         launch_arguments={
             'vehicle_config': LaunchConfiguration('vehicle_config'),
+            'frame_prefix': LaunchConfiguration('frame_prefix'),
+        }.items(),
+    )
+
+    waywiser_truck_localization_launch = IncludeLaunchDescription(
+        PythonLaunchDescriptionSource(
+            [
+                os.path.join(
+                    waywiser_core_dir,
+                    'launch',
+                    'waywiser_localization.launch.py',
+                )
+            ]
+        ),
+        launch_arguments={
+            'localization_config': LaunchConfiguration('vehicle_config'),
+            'localization_node_name': LaunchConfiguration('localization_node_name'),
         }.items(),
     )
 
@@ -100,11 +131,16 @@ def generate_launch_description():
         }.items(),
     )
 
+    # create opaque functions to launch nodes using context
+    urm14_ultrasonic_array_launch_action = OpaqueFunction(function=urm14_ultrasonic_array_launch)
+
     # create launch description
     ld = LaunchDescription()
 
     # declare launch args
     ld.add_action(vehicle_config_la)
+    ld.add_action(frame_prefix_la)
+    ld.add_action(localization_node_name_la)
     ld.add_action(enable_collision_monitor_la)
     ld.add_action(rviz_config_la)
     ld.add_action(teleop_config_la)
@@ -113,8 +149,48 @@ def generate_launch_description():
     ld.add_action(control_vehicle_node_name_la)
 
     # start nodes
-    ld.add_action(truck)
+    ld.add_action(waywiser_truck_launch)
+    ld.add_action(waywiser_truck_localization_launch)
     ld.add_action(twist_safety)
     ld.add_action(teleop_rviz2)
+    ld.add_action(urm14_ultrasonic_array_launch_action)
 
     return ld
+
+
+def urm14_ultrasonic_array_launch(context):
+    nodes = []
+
+    vehicle_config = get_full_file_path(LaunchConfiguration('vehicle_config').perform(context))
+    if vehicle_config == '':
+        return nodes
+
+    with open(vehicle_config, 'r', encoding='utf-8') as f:
+        config_data = yaml.safe_load(f)
+        node_params_dict = config_data['/**']['ros__parameters']
+        node_params_dict.update(config_data['waywiser_truck_node']['ros__parameters'])
+
+        if 'urm14_sensor_array_config' in node_params_dict:
+            urm14_sensor_array_config = get_full_file_path(
+                node_params_dict['urm14_sensor_array_config'],
+                os.path.join(get_package_share_directory('waywiser_hwbringup'), 'config'),
+            )
+
+            nodes.append(
+                IncludeLaunchDescription(
+                    PythonLaunchDescriptionSource(
+                        [
+                            os.path.join(
+                                get_package_share_directory('waywiser_hwbringup'),
+                                'launch',
+                                'urm14_ultrasonic_array.launch.py',
+                            )
+                        ]
+                    ),
+                    launch_arguments={
+                        'urm14_sensor_array_config': urm14_sensor_array_config,
+                    }.items(),
+                )
+            )
+
+    return nodes
